refactor: refactor and clean up imports & types
