+++ src/components/Footer.astro
@@ -1,6 +1,6 @@
 ---
 import { tinaField } from "@tinacms/astro/tina-field";
-import type { CmsConfig, CmsConfigContactLink } from '../lib/types/cms';
+import type { CmsConfig, CmsConfigContactLink } from '../lib/data';
 import IconLink from "./IconLink.astro";
 import Starfield from './space/Starfield.astro';
