+++ src/components/Header.astro
@@ -1,7 +1,7 @@
 ---
 import { Icon } from 'astro-icon/components';
 import { tinaField } from "@tinacms/astro/tina-field";
-import type { CmsConfig } from '../lib/types/cms';
+import type { CmsConfig } from '../lib/data';
 import IconLink from "./IconLink.astro";
 import ThemeToggle from './ThemeToggle.astro';
