+++ src/components/blocks/Blocks.astro
@@ -10,7 +10,7 @@ import Stats from './Stats.astro';
 import Testimonial from './Testimonial.astro';
 import Video from './Video.astro';
 import ConstellationDivider from '../space/ConstellationDivider.astro';
-import type { CmsPage } from '../../lib/types/cms';
+import type { CmsPage } from '../../lib/data';
 
 interface Props {
 	data?: CmsPage | null;
