+++ src/components/blocks/Callout.astro
@@ -2,7 +2,7 @@
 import { tinaField } from '@tinacms/astro/tina-field';
 import Section from '../ui/Section.astro';
 import Icon from '../ui/Icon.astro';
-import type { CalloutBlock } from '../../lib/types/cms';
+import type { CalloutBlock } from '../../lib/data';
 
 interface Props { data: CalloutBlock }
