+++ src/components/blocks/Content.astro
@@ -2,7 +2,7 @@
 import TinaMarkdown from '@tinacms/astro/TinaMarkdown.astro';
 import Section from '../ui/Section.astro';
 import { tinaField } from '@tinacms/astro/tina-field';
-import type { ContentBlock } from '../../lib/types/cms';
+import type { ContentBlock } from '../../lib/data';
 
 interface Props { data: ContentBlock }
