+++ src/components/blocks/Cta.astro
@@ -4,7 +4,7 @@ import Button from '../ui/Button.astro';
 import Icon from '../ui/Icon.astro';
 import { tinaField } from '@tinacms/astro/tina-field';
 import OrbitRings from '../space/OrbitRings.astro';
-import type { CtaBlock } from '../../lib/types/cms';
+import type { CtaBlock } from '../../lib/data';
 
 interface Props { data: CtaBlock }
