+++ src/components/blocks/Features.astro
@@ -6,7 +6,7 @@ import Card from '../ui/Card.astro';
 import CardHeader from '../ui/CardHeader.astro';
 import CardContent from '../ui/CardContent.astro';
 import Icon from '../ui/Icon.astro';
-import type { FeaturesBlock } from '../../lib/types/cms';
+import type { FeaturesBlock } from '../../lib/data';
 
 interface Props { data: FeaturesBlock }
