+++ src/components/blocks/Hero.astro
@@ -5,7 +5,7 @@ import Button from '../ui/Button.astro';
 import Icon from '../ui/Icon.astro';
 import Aurora from '../space/Aurora.astro';
 import Starfield from '../space/Starfield.astro';
-import type { HeroBlock } from '../../lib/types/cms';
+import type { HeroBlock } from '../../lib/data';
 
 interface Props { data: HeroBlock }
