+++ src/components/blocks/Split.astro
@@ -5,7 +5,7 @@ import Section from '../ui/Section.astro';
 import Button from '../ui/Button.astro';
 import Icon from '../ui/Icon.astro';
 import { cn } from '../../lib/cn';
-import type { SplitBlock } from '../../lib/types/cms';
+import type { SplitBlock } from '../../lib/data';
 
 interface Props { data: SplitBlock }
