+++ src/components/blocks/Stats.astro
@@ -2,7 +2,7 @@
 import { tinaField } from '@tinacms/astro/tina-field';
 import Section from '../ui/Section.astro';
 import Planet from '../space/Planet.astro';
-import type { StatsBlock } from '../../lib/types/cms';
+import type { StatsBlock } from '../../lib/data';
 
 interface Props { data: StatsBlock }
