+++ src/components/blocks/Testimonial.astro
@@ -4,7 +4,7 @@ import Section from '../ui/Section.astro';
 import Card from '../ui/Card.astro';
 import CardContent from '../ui/CardContent.astro';
 import Avatar from '../ui/Avatar.astro';
-import type { TestimonialBlock, TestimonialItem } from '../../lib/types/cms';
+import type { TestimonialBlock, TestimonialItem } from '../../lib/data';
 
 interface Props { data: TestimonialBlock }
