+++ src/components/blocks/Video.astro
@@ -1,7 +1,7 @@
 ---
 import { tinaField } from '@tinacms/astro/tina-field';
 import Section from '../ui/Section.astro';
-import type { VideoBlock } from '../../lib/types/cms';
+import type { VideoBlock } from '../../lib/data';
 
 interface Props { data: VideoBlock }
