+++ src/components/islands/BlogBody.astro
@@ -1,7 +1,7 @@
 ---
 import TinaMarkdown from '@tinacms/astro/TinaMarkdown.astro';
 import { tinaField } from '@tinacms/astro/tina-field';
-import type { CmsBlog } from '../../lib/types/cms';
+import type { CmsBlog } from '../../lib/data';
 import FormattedDate from '../FormattedDate.astro';
 
 interface Props {
