+++ src/components/islands/PageBody.astro
@@ -1,6 +1,6 @@
 ---
 import Blocks from '../blocks/Blocks.astro';
-import type { CmsPage } from '../../lib/types/cms';
+import type { CmsPage } from '../../lib/data';
 
 interface Props {
 	data?: CmsPage | null;
