+++ src/lib/data.ts
@@ -1,48 +1,38 @@
 /**
- * Per-collection data loaders. Each calls the generated Tina client the
- * normal way, then pipes the result through `requestWithMetadata()` so the
- * editor overlay flows in automatically when the page is rendered inside
- * the admin iframe, and `tinaField()` has the metadata it needs.
+ * Per-collection data loaders + the data shapes they return.
  *
- * Return shapes are widened/narrowed via the wrapper types in
- * `./types/cms` so consumers don't have to know about the generated-vs-
- * actual schema drift.
+ * Loaders call the generated Tina client and pipe the result through
+ * `requestWithMetadata()` so the editor overlay flows in when the page
+ * renders inside the admin iframe and `tinaField()` has its metadata.
+ *
+ * Types below are pure derivations — no hand-written shapes. Each one is
+ * either inferred from a loader's return type (`CmsConfig`/`CmsPage`/
+ * `CmsBlog`) or `Extract`/index-accessed off those. The Tina collection
+ * is the source of truth; regen with `tinacms dev` and everything
+ * downstream updates.
  */
+import type { TinaRichTextContent } from '@tinacms/astro';
 import { requestWithMetadata } from '@tinacms/astro/data';
 import client from '../../tina/__generated__/client';
-import type { CmsBlog, CmsConfig, CmsPage } from './types/cms';
 
 export const getConfig = () =>
-	requestWithMetadata(client.queries.config({ relativePath: 'config.json' })) as Promise<{
-		data?: { config?: CmsConfig | null } | null;
-	}>;
+	requestWithMetadata(client.queries.config({ relativePath: 'config.json' }));
 
 export const getPage = (slug: string) =>
-	requestWithMetadata(
-		client.queries.page({ relativePath: `${slug}.mdx` }),
-		{ priority: 'primary' },
-	) as Promise<{ data?: { page?: CmsPage | null } | null }>;
+	requestWithMetadata(client.queries.page({ relativePath: `${slug}.mdx` }), { priority: 'primary' });
 
 export const getBlog = (slug: string) =>
-	requestWithMetadata(
-		client.queries.blog({ relativePath: `${slug}.mdx` }),
-		{ priority: 'primary' },
-	) as Promise<{ data?: { blog?: CmsBlog | null } | null }>;
+	requestWithMetadata(client.queries.blog({ relativePath: `${slug}.mdx` }), { priority: 'primary' });
 
-export async function listPages(): Promise<CmsPage[]> {
+export async function listPages() {
 	const result = await client.queries.pageConnection();
-	return (result?.data?.pageConnection?.edges ?? [])
-		.flatMap((edge) => (edge?.node ? [edge.node as unknown as CmsPage] : []));
+	return (result.data.pageConnection.edges ?? [])
+		.flatMap((edge) => (edge?.node ? [edge.node] : []));
 }
 
-/** Derive a blog post's URL slug from its source file path. */
-export function blogSlug(node: { _sys?: { relativePath?: string | null } | null }): string {
-	return node._sys?.relativePath?.replace(/\.mdx?$/, '') ?? '';
-}
-
-export async function listBlogs(): Promise<CmsBlog[]> {
+export async function listBlogs() {
 	const result = await client.queries.blogConnection();
-	return (result?.data?.blogConnection?.edges ?? [])
+	return (result.data.blogConnection.edges ?? [])
 		.flatMap((edge) => (edge?.node ? [edge.node] : []))
 		.sort((a, b) => {
 			const ad = a.pubDate ? new Date(a.pubDate).valueOf() : 0;
@@ -50,3 +40,33 @@ export async function listBlogs(): Promise<CmsBlog[]> {
 			return bd - ad;
 		});
 }
+
+export type CmsConfig = Awaited<ReturnType<typeof getConfig>>['data']['config'];
+export type CmsPage = Awaited<ReturnType<typeof getPage>>['data']['page'];
+export type CmsBlog = Awaited<ReturnType<typeof getBlog>>['data']['blog'];
+
+export type PageBlock = NonNullable<NonNullable<CmsPage['blocks']>[number]>;
+export type PageBlockTypename = PageBlock['__typename'];
+
+export type HeroBlock = Extract<PageBlock, { __typename: 'PageBlocksHero' }>;
+export type CalloutBlock = Extract<PageBlock, { __typename: 'PageBlocksCallout' }>;
+export type FeaturesBlock = Extract<PageBlock, { __typename: 'PageBlocksFeatures' }>;
+export type StatsBlock = Extract<PageBlock, { __typename: 'PageBlocksStats' }>;
+export type CtaBlock = Extract<PageBlock, { __typename: 'PageBlocksCta' }>;
+export type ContentBlock = Extract<PageBlock, { __typename: 'PageBlocksContent' }>;
+export type TestimonialBlock = Extract<PageBlock, { __typename: 'PageBlocksTestimonial' }>;
+export type VideoBlock = Extract<PageBlock, { __typename: 'PageBlocksVideo' }>;
+export type SplitBlock = Extract<PageBlock, { __typename: 'PageBlocksSplit' }>;
+
+export type CmsConfigNav = NonNullable<NonNullable<CmsConfig['nav']>[number]>;
+export type CmsConfigContactLink = NonNullable<NonNullable<CmsConfig['contactLinks']>[number]>;
+export type CmsConfigSeo = NonNullable<CmsConfig['seo']>;
+
+export type Action = NonNullable<NonNullable<HeroBlock['actions']>[number]>;
+export type ImageField = NonNullable<HeroBlock['image']>;
+export type FeatureItem = NonNullable<NonNullable<FeaturesBlock['items']>[number]>;
+export type StatItem = NonNullable<NonNullable<StatsBlock['stats']>[number]>;
+export type TestimonialItem = NonNullable<NonNullable<TestimonialBlock['testimonials']>[number]>;
+
+/** Tina rich-text bodies are typed as `any` in the generated client; this is what `<TinaMarkdown>` expects. */
+export type RichText = TinaRichTextContent;
