+++ src/pages/blog/[...slug].astro
@@ -7,10 +7,7 @@ import { islands } from '../../lib/islands';
 
 export async function getStaticPaths() {
 	const posts = await listBlogs();
-	return posts
-		.map((node) => node._sys?.relativePath?.replace(/\.mdx?$/, '') ?? '')
-		.filter(Boolean)
-		.map((slug) => ({ params: { slug } }));
+	return posts.map((node) => ({ params: { slug: node._sys.filename } }));
 }
 
 const slug = (Astro.params.slug ?? '').toString();
