+++ src/pages/blog/index.astro
@@ -1,11 +1,9 @@
 ---
 import Base from '../../layouts/Base.astro';
 import FormattedDate from '../../components/FormattedDate.astro';
-import { blogSlug, getConfig, listBlogs } from '../../lib/data';
+import { getConfig, listBlogs } from '../../lib/data';
 
-const posts = (await listBlogs())
-	.map((post) => ({ post, slug: blogSlug(post) }))
-	.filter(({ slug }) => slug);
+const posts = await listBlogs();
 const config = (await getConfig()).data?.config ?? null;
 ---
 
@@ -16,9 +14,9 @@ const config = (await getConfig()).data?.config ?? null;
 	<main class="mx-auto max-w-5xl px-6 py-12">
 		<h1 class="mb-8 text-4xl font-semibold tracking-tight">Blog</h1>
 		<ul class="grid list-none gap-6 p-0 sm:grid-cols-2">
-			{posts.map(({ post, slug }) => (
+			{posts.map((post) => (
 				<li>
-					<a href={`/blog/${slug}/`} class="group block overflow-hidden rounded-xl border bg-card shadow-sm transition-shadow hover:shadow-md">
+					<a href={`/blog/${post._sys.filename}/`} class="group block overflow-hidden rounded-xl border bg-card shadow-sm transition-shadow hover:shadow-md">
 						{post.heroImage && (
 							<img width={720} height={360} src={post.heroImage} alt={post.title ?? ''} class="aspect-video w-full object-cover" />
 						)}
